requirements.txt
streamlit==1.43.2
google-generativeai==0.3.2
python-dotenv==1.0.1
requests==2.31.0